# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Rathinam Crackers Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/transfers:
    get:
      operationId: listTransfers
      tags: [transfers]
      summary: List stock transfers
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: fromLocationId
          in: query
          schema:
            type: string
        - name: toLocationId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Transfer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferListResponse"
    post:
      operationId: createTransfer
      tags: [transfers]
      summary: Create stock transfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferBody"
      responses:
        "201":
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /v1/transfers/pending:
    get:
      operationId: listPendingTransfers
      tags: [transfers]
      summary: List pending and in-transit transfers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending transfers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferListResponse"
  /v1/transfers/{id}:
    get:
      operationId: getTransfer
      tags: [transfers]
      summary: Get transfer by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
  /v1/transfers/{id}/dispatch:
    put:
      operationId: dispatchTransfer
      tags: [transfers]
      summary: Dispatch stock transfer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchTransferBody"
      responses:
        "200":
          description: Transfer dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /v1/transfers/{id}/receive:
    put:
      operationId: receiveTransfer
      tags: [transfers]
      summary: Receive stock transfer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveTransferBody"
      responses:
        "200":
          description: Transfer received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # ============================================================
  # COUPONS
  # ============================================================
...
  /v1/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /v1/users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ============================================================
  # LOCATIONS
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # AUTH SCHEMAS
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, ERP_MANAGER, ACCOUNTANT, AGENT, WH_MANAGER, CASHIER]
        locationIds:
          type: array
          items:
            type: string
        email:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CreateUserBody:
      type: object
      required: [name, username, password, role]
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        pin:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        locationIds:
          type: array
          items:
            type: string
        maxDiscountPct:
          type: number
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        locationIds:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        maxDiscountPct:
          type: number

    # PRODUCT SCHEMAS
...
    TransferItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        variantId:
          type: string
        qty:
          type: integer
        receivedQty:
          type: integer
        batchNo:
          type: string
    Transfer:
      type: object
      properties:
        id:
          type: string
        transferNo:
          type: string
        fromLocationId:
          type: string
        fromLocationName:
          type: string
        toLocationId:
          type: string
        toLocationName:
          type: string
        status:
          type: string
          enum: [draft, pending_approval, in_transit, received, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/TransferItem"
        vehicleNo:
          type: string
        notes:
          type: string
        dispatchedAt:
          type: string
        receivedAt:
          type: string
        createdAt:
          type: string
    TransferListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transfer"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    TransferResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Transfer"
    CreateTransferBody:
      type: object
      required: [fromLocationId, toLocationId, items]
      properties:
        fromLocationId:
          type: string
        toLocationId:
          type: string
        items:
          type: array
          items:
            type: object
            required: [productId, variantId, qty]
            properties:
              productId:
                type: string
              variantId:
                type: string
              qty:
                type: integer
              batchNo:
                type: string
        notes:
          type: string
    DispatchTransferBody:
      type: object
      properties:
        vehicleNo:
          type: string
        notes:
          type: string
    ReceiveTransferBody:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              variantId:
                type: string
              receivedQty:
                type: integer

    # COUPON SCHEMAS
...